requirements.txt
streamlit
tensorflow-cpu==2.10.0
numpy
opencv-python
Pillow
